pyproject.toml
[tool.poetry]
name = "equipment_cyg"
version = "1.0.0"
description = ""
repository = "https://github.com/anjiyou/equipment_cyg"
homepage = "https://github.com/anjiyou/equipment_cyg"
authors = ["WeiLiu <wei.l@cyg.com>"]
readme = "README.md"
packages = [{include = "equipment_cyg"},]


[tool.poetry.dependencies]
python = "^3.10"
secsgem_cyg="1.0.0"
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
modbus-tk = "^1.1.3"
isort = "^5.13.2"
flake8 = "^7.1.0"


[tool.poetry.group.dev.dependencies]
secsgem_cyg = { git = "https://github.com/anjiyou/secsgem_cyg.git", branch = "main" }


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
